refactor: move transaction hash functions to separate file
